docker_manager/docker_service_manager.py
import subprocess
import docker
from docker.errors import APIError

class DockerServiceManager:

    @staticmethod
    def is_docker_running():
        """Internal method to check if the Docker service is running."""
        client = docker.from_env()
        try:
            client.ping()  # pings the Docker daemon
            return True
        except APIError:
            return False
        finally:
            client.close()

    @staticmethod
    def start_docker():
        """Internal method to start the Docker service."""
        # The Docker SDK for Python doesn't provide a direct method to start the Docker service.
        # This operation typically requires system-level permissions and is usually done outside
        # the scope of a Docker client application.
        result = subprocess.run(['sudo', 'service', 'docker', 'start'], stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
        return result.returncode == 0
